+++ integrations/ragie/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @mastra/ragie
 
+## 1.0.1-alpha.45
+
+### Patch Changes
+
+- Updated dependencies [14064f2]
+  - @mastra/core@0.1.27-alpha.66
+
+## 1.0.1-alpha.44
+
+### Patch Changes
+
+- Updated dependencies [e66643a]
+  - @mastra/core@0.1.27-alpha.65
+
 ## 1.0.1-alpha.43
 
 ### Patch Changes